DTO 수정

+++ Angbit/src/main/java/com/oracle/Angbit/model/common/Trade.java
@@ -13,7 +13,7 @@ public class Trade {
 	private String id;
 	private String coincode;
 	private Date trd_date;
-	private int trd_amt;
+	private Float trd_amt;
 	private int trd_unit_price;
 	private int trd_price;
 	private int trd_div;
